+++ api/v1alpha1/groupversion_info.go
@@ -20,8 +20,9 @@ limitations under the License.
 package v1alpha1
 
 import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"sigs.k8s.io/controller-runtime/pkg/scheme"
 )
 
 var (
@@ -29,8 +30,20 @@ var (
 	GroupVersion = schema.GroupVersion{Group: "policy.giantswarm.io", Version: "v1alpha1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
-	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
+	SchemeBuilder = &runtime.SchemeBuilder{}
 
 	// AddToScheme adds the types in this group-version to the given scheme.
 	AddToScheme = SchemeBuilder.AddToScheme
 )
+
+// addKnownTypes registers objs with the scheme under GroupVersion. It replaces
+// controller-runtime's scheme.Builder, which is deprecated because api packages
+// should depend only on the standard library, apimachinery and other api
+// packages.
+func addKnownTypes(objs ...runtime.Object) {
+	SchemeBuilder.Register(func(s *runtime.Scheme) error {
+		s.AddKnownTypes(GroupVersion, objs...)
+		metav1.AddToGroupVersion(s, GroupVersion)
+		return nil
+	})
+}